src/requirements.txt
appdirs==1.4.4
autopep8==1.4.4
bleach==3.1.5
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.8
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
docutils==0.16
filelock==3.0.12
idna==2.9
jsmin==2.2.2
keyring==21.2.1
oauthlib==3.1.0
packaging==20.4
pep8==1.7.1
Pillow==7.1.2
pipenv==2020.5.28
pkginfo==1.5.0.1
pycodestyle==2.5.0
Pygments==2.6.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2018.5
pywin32-ctypes==0.2.0
readme-renderer==26.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.10.14
six==1.15.0
sqlparse==0.2.4
tqdm==4.46.0
twine==3.1.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
webencodings==0.5.1
